orin_refactor/requirements.txt
aiofiles
aiohttp
aiosqlite
ffmpeg_python
numpy
opencv_python
psutil
python-dotenv
scipy
ultralytics
websockets